Added JSDocs moved some variables from functions.ts to constants.ts made some variables private added getters to moved variables

+++ src/commands.ts
@@ -6,28 +6,33 @@
 
 import { closeIcon, addIcon, redoIcon } from '@jupyterlab/ui-components';
 import { FileDialog } from '@jupyterlab/filebrowser';
-import { InputDialog } from '@jupyterlab/apputils';
+import { InputDialog, showErrorMessage } from '@jupyterlab/apputils';
 
 // Custom imports
 import {
   commandPrefix,
   bookmarkLaunchers,
   categories,
-  launcher,
-  docManager,
-  commands,
-  notebookTracker,
-  FAVORITE_ICON
+  FAVORITE_ICON,
+  UNCATEGORIZED,
+  getBookmarks,
+  getNotebookTracker,
+  getCommands,
+  getLauncher,
+  getDocManager
 } from './constants';
 import {
   syncBookmark,
   addBookmarkItem,
   deleteBookmark,
   addCategory,
-  deleteCategory,
-  bookmarks
+  deleteCategory
 } from './functions';
 
+/**
+ * This command is used when the user opens the context menu (makes a right click)
+ * of an open notebook, and selects `Add bookmark to...` menu.
+ */
 export const addBookmarkContextMenuCommand = {
   id: commandPrefix + 'addBookmark',
   options: {
@@ -40,24 +45,33 @@ export const addBookmarkContextMenuCommand = {
           return item[0];
         })
       }).then(result => {
-        const category =
-          result.value === undefined ? 'Uncategorized' : result.value;
-        const currentDoc = notebookTracker.currentWidget;
-        currentDoc.context.fileChanged.connect(syncBookmark);
-        const currentDocName = currentDoc.context.contentsModel.name;
-        const currentDocPath = currentDoc.context.path;
-        addBookmarkItem(
-          commands,
-          launcher,
-          currentDocName,
-          currentDocPath,
-          category
-        );
+        if (result.button.label !== 'Cancel') {
+          const category =
+            result.value === undefined ? 'Uncategorized' : result.value;
+          const currentDoc = getNotebookTracker().currentWidget;
+          currentDoc.context.fileChanged.connect(syncBookmark);
+          const currentDocName = currentDoc.context.contentsModel.name;
+          const currentDocPath = currentDoc.context.path;
+          addBookmarkItem(
+            getCommands(),
+            getLauncher(),
+            currentDocName,
+            currentDocPath,
+            category
+          );
+        }
       });
     }
   }
 };
 
+/**
+ * This command is used by the `Launcher's` `Add Bookmark` items.
+ * The item in the `Management` category let's the user select a category to add the bookmark to.
+ * The items in the custom (user defined) categories always provide the category they belong to via the `arg`
+ * JSON object.
+ * The bookmark is then added to the selected category.
+ */
 export const addBookmarkLauncherCommand = {
   id: commandPrefix + 'addBookmarkFromLauncher',
   options: {
@@ -65,26 +79,60 @@ export const addBookmarkLauncherCommand = {
     caption: 'Add bookmark',
     icon: FAVORITE_ICON,
     execute: (args: any): void => {
-      FileDialog.getOpenFiles({
-        manager: docManager,
-        filter: model => model.type === 'notebook'
-      }).then(result => {
-        if (result.button.label !== 'Cancel') {
-          result.value.forEach(selectedFile => {
-            addBookmarkItem(
-              commands,
-              launcher,
-              selectedFile.name,
-              selectedFile.path,
-              args.category
-            );
-          });
-        }
-      });
+      if (args.category) {
+        FileDialog.getOpenFiles({
+          manager: getDocManager(),
+          filter: model => model.type === 'notebook'
+        }).then(result => {
+          if (result.button.label !== 'Cancel') {
+            result.value.forEach(selectedFile => {
+              addBookmarkItem(
+                getCommands(),
+                getLauncher(),
+                selectedFile.name,
+                selectedFile.path,
+                args.category
+              );
+            });
+          }
+        });
+      } else {
+        InputDialog.getItem({
+          title: 'Select category',
+          items: Array.from(categories, item => {
+            return item[0];
+          })
+        }).then(result => {
+          if (result.button.label !== 'Cancel') {
+            const category =
+              result.value === undefined ? 'Uncategorized' : result.value;
+            FileDialog.getOpenFiles({
+              manager: getDocManager(),
+              filter: model => model.type === 'notebook'
+            }).then(result => {
+              if (result.button.label !== 'Cancel') {
+                result.value.forEach(selectedFile => {
+                  addBookmarkItem(
+                    getCommands(),
+                    getLauncher(),
+                    selectedFile.name,
+                    selectedFile.path,
+                    category
+                  );
+                });
+              }
+            });
+          }
+        });
+      }
     }
   }
 };
-
+/**
+ * This command is used to remove a bookmark. The argument of the command is the category
+ * to remove the bookmark from. The command only lists bookmarks for removal that belong to that
+ * category. If the command is invoked from the `Management` section, then all bookmarks are listed.
+ */
 export const removeBookmarkCommand = {
   id: commandPrefix + 'removeBookmark',
   options: {
@@ -99,7 +147,7 @@ export const removeBookmarkCommand = {
             ? Array.from(bookmarkLaunchers, item => {
                 return item[0];
               })
-            : Array.from(bookmarks, entry => {
+            : Array.from(getBookmarks(), entry => {
                 if (entry[1].category === args.category) {
                   return entry[1].title;
                 }
@@ -114,6 +162,12 @@ export const removeBookmarkCommand = {
   }
 };
 
+/**
+ * Adds a category to the launcher. Empty names are not allowed
+ * and eventually duplicates are refused to be added.
+ * Newly added categories are initialized with the two built-in command:
+ * add and delete bookmark.
+ */
 export const addCategoryCommand = {
   id: commandPrefix + 'addCategory',
   options: {
@@ -126,13 +180,29 @@ export const addCategoryCommand = {
       }).then(result => {
         if (result.button.label !== 'Cancel') {
           const categoryToAdd: string = result.value;
-          addCategory(categoryToAdd);
+          if (
+            result.value === null ||
+            result.value === '' ||
+            result.value === undefined
+          ) {
+            showErrorMessage(
+              'Invalid category name',
+              'Category name cannot be empty.'
+            );
+          } else {
+            addCategory(categoryToAdd);
+          }
         }
       });
     }
   }
 };
 
+/**
+ * Deletes a category. The `Uncategorized` category cannot be deleted.
+ * Eventually the bookmarks from the deleted category will be moved to the `Uncategorized`
+ * section.
+ */
 export const deleteCategoryCommand = {
   id: commandPrefix + 'deleteCategory',
   options: {
@@ -148,13 +218,23 @@ export const deleteCategoryCommand = {
       }).then(result => {
         if (result.button.label !== 'Cancel') {
           const categoryToDelete: string = result.value;
-          deleteCategory(categoryToDelete);
+          if (categoryToDelete === UNCATEGORIZED) {
+            showErrorMessage(
+              'Category cannot be deleted',
+              `The default "${UNCATEGORIZED}" category cannot be deleted.`
+            );
+          } else {
+            deleteCategory(categoryToDelete);
+          }
         }
       });
     }
   }
 };
 
+/**
+ * Not used for the moment.
+ */
 export const moveToCategoryCommand = {
   id: commandPrefix + 'moveToCategory',
   options: {
